Рефакторинг исключения
